pyproject.toml
[tool.poetry]
name = "timetable"
version = "0.1.0"
description = "Skvortcov Aleksei"
authors = ["Skvortcov Aleksei"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]

python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
pydantic = {extras = ["email"], version = "^2.7.2"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.3"
python-multipart = "^0.0.9"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
celery = "^5.4.0"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
  ".", "timetable",
]
asyncio_mode="auto"
